+++ src/engine/buffer.cpp
@@ -22,11 +22,11 @@ namespace Voxel {
         glBufferData(GL_ARRAY_BUFFER, data_size, data, GL_STATIC_DRAW);
     }
 
-    void VBO::mapped_data(float* data, size_t data_size)
+    void VBO::mapped_data(uint32_t* data, size_t data_size)
     {
         GLbitfield flags = GL_MAP_WRITE_BIT | GL_MAP_PERSISTENT_BIT | GL_MAP_COHERENT_BIT;
         glBufferStorage(GL_ARRAY_BUFFER, data_size, data, flags);
-        buffer_memory_ptr = (float*)glMapBufferRange(GL_ARRAY_BUFFER, 0, data_size, flags);
+        buffer_memory_ptr = (uint32_t*)glMapBufferRange(GL_ARRAY_BUFFER, 0, data_size, flags);
     }
 
     EBO::EBO() {
